Caesar Cipher
# Assignment: Caesar Cipher
For this assignment I had to write a program, named caesar_cipher.py, that will prompt the end-user for the plaintext to be encrypted and the shift factor. Then create a ciphertext string by shifting the input by that many characters to the right. Display the encrypted string to the user.
Finally, wrote another program, named caesar_decrypt.py, that will take an encrypted string and factor as input and decrypt the string by subtracting the factor from each character's ASCII code.