Guess
# Assignment: Guess

Write a program, named guess, that asks the user to guess a number between 1 and 100. The program should end when the user guesses correctly, but otherwise provides hints for guessing higher or lower.

Provide a user-friendly reminder if the end-user does not enter a number as required.

Extra credit: After each game is over, ask if the user would like to play again. Allow the user to play as many times as they would like. Use multiple functions to separate the playing of a single game from the repetition query.